+++ application/models/SalesHistory.php
@@ -2,6 +2,6 @@
 class salesHistory extends MY_Model {
         public function __construct()
         {
-                parent::__construct('salesHistory', 'timeofTransaction');
+                parent::__construct('salesHistory', 'timestamp');
         }
 }