+++ application/views/itemnav.php
@@ -1,0 +1,10 @@
+<div style="margin-left:2%">
+    <div class="col-xs-12 list">
+        <ul class="nav nav-pills">
+            <li><a class="btn btn-primary" href="/viewhistory/page/{first}">First</a></li>
+            <li><a class="btn btn-primary" href="/viewhistory/page/{previous}">Previous</a></li>
+            <li><a class="btn btn-primary" href="/viewhistory/page/{next}">Next</a></li>
+            <li><a class="btn btn-primary" href="/viewhistory/page/{last}">Last</a></li>
+        </ul>
+    </div>
+</div>